Complete Upgrade
- code organization
- add curve parameter to manage transfer function’ parameter
- add value listener
- etc.
