+++ knob.kv
@@ -1,0 +1,81 @@
+#
+#    Knob
+#    ====
+#     To create a basic knob (in a kv file):
+#
+#     Knob:
+#       size:               100, 100
+#       min:                0
+#       max:                100
+#       step:               1
+#       slope:              1
+#       value:              0                       # Default position of knob.
+#       knobimg_source:     "img/knob_metal.png"    # Knob texture
+#       show_label:         True                    # Show central label
+#       show_marker:        False                   # Do not show surrounding marker
+#
+#     To create a knob with a surrounding marker:
+#
+#     Knob:
+#       size:               100, 100
+#       min:                0
+#       max:                100
+#       step:               1
+#       slope:              1
+#       value:              0                       # Default position of knob.
+#       knobimg_source:     "img/knob_metal.png"    # Knob texture
+#       show_label:         True                    # Show central label
+#       show_marker:        True                    # Show surrounding marker
+#       marker_img:         "img/bline.png"         # Marker texture image
+#       knob_size:          0.9                     # Scales knob size to leave space for marker
+#       markeroff_color:    0, 0, 0, 0
+
+<Knob>
+    label: _label
+    size_hint: None, None
+
+    canvas.before:
+        Color:
+            rgba: self.markeroff_color
+        Ellipse:
+            pos: self.pos
+            size: self.size[0], self.size[1]
+            angle_start: 0
+            angle_end: 360
+            source: self.markeroff_img
+
+        Color:
+            rgba: self.marker_color
+        Ellipse:
+            pos: self.pos
+            size: self.size[0], self.size[1]
+            angle_start: self.marker_startangle
+            angle_end: self._angle + self.marker_ahead if self._angle > self.marker_startangle else self.marker_startangle
+            source: self.marker_img
+
+        Color:
+            rgba: self.knobimg_bgcolor
+        Ellipse:
+            pos: self.pos[0] + (self.size[0] * (1 - self.knobimg_size))/2, self.pos[1] + (self.size[1] * (1 - self.knobimg_size)) / 2
+            size: self.size[0] * (self.knobimg_size), self.size[1] * (self.knobimg_size)
+
+        Color:
+            rgba: self.knobimg_color
+        PushMatrix
+        Rotate:
+            angle: 360 - self._angle
+            origin: self.center
+        Rectangle:
+            pos: self.pos[0] + (self.size[0] * (1 - self.knobimg_size)) /2, self.pos[1] + (self.size[1] * (1 - self.knobimg_size)) / 2
+            size: self.size[0] * (self.knobimg_size), self.size[1] * (self.knobimg_size)
+            source: self.knobimg_source
+
+    canvas:
+        PopMatrix
+
+    Label:
+        id: _label
+        text: "%.2f"%(root.value)
+        center: root.center
+        font_size: root.font_size
+        color: root.font_color
